pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "rnaflow"
version = "0.1.0"
description = "Flow-inspired latent space optimization of mRNA sequences for cell-type specificity"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy",
    "pyyaml",
    "tqdm",
    "matplotlib",
    "pytorch-lightning>=2.0",
    "pandas",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[tool.setuptools.packages.find]
include = ["rnaflow*"]
